requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.3.0
python-decouple==3.8
reportlab==4.4.3
requests==2.32.5
celery==5.5.2
redis==6.4.0
python-dateutil==2.9.0.post0
PyJWT==2.10.1
psycopg2-binary==2.9.10
whitenoise==6.9.0
gunicorn==23.0.0
